products/tier3/ocr-processamento/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis e cache
redis==5.0.1
hiredis==2.2.3

# Celery (processamento assíncrono)
celery==5.3.4
flower==2.0.1

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Google Vision API
google-cloud-vision==3.4.5

# OpenAI API
openai==1.3.5

# Processamento de texto
python-dateutil==2.8.2
langchain==0.0.335

# Utilitários
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1

# Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging e monitoramento
structlog==23.2.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Outros
numpy==1.26.2

